+++ packages/tailwind-helpers/src/tw-color-palette/index.ts
@@ -1,0 +1,1 @@
+export { twColorPalette } from "./tw-color-palette";
