+++ packages/tailwind-helpers/src/tw-pseudo/index.ts
@@ -1,0 +1,1 @@
+export { twPseudo } from "./tw-pseudo";
