--- packages/tailwind-helpers/src/twColorPalette/index.ts
@@ -1,1 +1,0 @@
-export { twColorPalette } from "./twColorPalette";
