--- packages/tailwind-helpers/src/twPseudo/index.ts
@@ -1,1 +1,0 @@
-export { twPseudo } from "./twPseudo";
